requirements.txt
scipy
matplotlib
tqdm
numpy
opencv-python
pillow
tensorboard
torch
torchvision